refactor(tauri): split AppState and the agent turn pipeline out of lib.rs
lib.rs mixed command registration, app state, and the whole turn flow
(7.9k lines), so state-shape and turn-flow changes kept churning the
registration hub. Pure code motion, no behavior change:

- src/app_state.rs: AppState, SessionRuntime, ActiveProject,
  ProjectActivityLocks, and the MCP App context helpers.
- src/agent_turn.rs: send_message and its inner driver, the parked
  turn queue (enqueue/cut-in/driver), and stop_agent.

pub(crate) re-exports keep every existing crate::X path working;
generate_handler! now uses agent_turn::-qualified paths, matching the
existing session_commands pattern.
